requirements.txt
Django==4.2
djangorestframework==3.15.1
aiogram==3.3.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-modeltranslation==0.19.3
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
setuptools==70.2.0
pillow==10.4.0
django-environ==0.11.2